+++ R/wget.R
@@ -20,7 +20,7 @@ wget <- function(input_url,
     # https://stackoverflow.com/questions/21365251/how-to-run-wget-in-background-for-an-unattended-download-of-files
     ## -bqc makes wget run in the background quietly
     dir.create(dirname(output_path), showWarnings = FALSE, recursive = TRUE) 
-    messager("Downloading with wget",paste0("[",1,"thread]:"),"\n",
+    messager("Downloading with wget",paste0("[",1," thread]:"),"\n",
              input_url,"==>",output_path, v=verbose) 
     cmd <- paste(
         wget_path,
@@ -35,7 +35,6 @@ wget <- function(input_url,
         "-O", output_path,
         if (force_overwrite) "" else "--no-clobber"
     ) 
-    system(cmd)
-    messager("\nwget download complete.",v=verbose)
+    system(cmd) 
     return(output_path)
 }